# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoMindscout API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/{bookingId}:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Get chat messages for a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a booking chat
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"

  # ── Disputes ───────────────────────────────────────────────────────────────
  /disputes:
    post:
      operationId: createDispute
      tags: [disputes]
      summary: Open a dispute for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Created dispute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{bookingId}:
    get:
      operationId: getDispute
      tags: [disputes]
      summary: Get dispute for a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Dispute details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"

  # ── Payouts ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ChatMessage:
      type: object
      required: [id, bookingId, senderId, content, isFlagged, createdAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        senderId: { type: integer }
        senderName: { type: ["string", "null"] }
        senderAvatarUrl: { type: ["string", "null"] }
        content: { type: string }
        isFlagged: { type: boolean }
        flagReason: { type: ["string", "null"] }
        createdAt: { type: string }
    ChatMessageResponse:
      type: object
      required: [id, bookingId, senderId, content, isFlagged, createdAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        senderId: { type: integer }
        senderName: { type: ["string", "null"] }
        senderAvatarUrl: { type: ["string", "null"] }
        content: { type: string }
        isFlagged: { type: boolean }
        flagReason: { type: ["string", "null"] }
        warning: { type: ["string", "null"] }
        createdAt: { type: string }
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    Dispute:
      type: object
      required: [id, bookingId, openedByUserId, reason, description, status, createdAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        openedByUserId: { type: integer }
        openerName: { type: ["string", "null"] }
        reason: { type: string }
        description: { type: string }
        evidenceUrl: { type: ["string", "null"] }
        status: { type: string, enum: [open, under_review, resolved] }
        adminDecision: { type: ["string", "null"] }
        resolutionType: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    DisputeInput:
      type: object
      required: [bookingId, reason, description]
      properties:
        bookingId: { type: integer }
        reason:
          type: string
          enum: [mentor_no_show, mentee_no_show, technical_issue, wrong_expertise, misconduct, other]
        description: { type: string }
        evidenceUrl: { type: string }
...